# back/openapi.yaml
openapi: 3.0.3
info:
  title: HRMS Dynamic Enterprise Platform API
  version: 1.0.0
  description: A configurable Human Resource Management System where administrators
    can create and modify business entities entirely from the UI.
  contact:
    name: HRMS API Team
    email: api-team@example.com
  license:
    name: Proprietary
    url: https://example.com/license
paths:
  /api/auth/login:
    post:
      operationId: auth_login
      summary: Authenticate user and issue JWT tokens
      tags:
      - Authentication
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              LoginRequest:
                value:
                  username: admin
                  password: password
                summary: Login Request
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/LoginRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccessEnvelope'
              examples:
                LoginResponse:
                  value:
                    success: true
                    data:
                      access: <jwt-access-token>
                      refresh: <jwt-refresh-token>
                      user:
                        id: 1
                        username: admin
                  summary: Login Response
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
  /api/auth/logout:
    post:
      operationId: auth_logout
      summary: Blacklist refresh token and logout
      tags:
      - Authentication
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
            examples:
              LogoutRequest:
                value:
                  refresh: <jwt-refresh-token>
                summary: Logout Request
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccessEnvelope'
              examples:
                LogoutResponse:
                  value:
                    success: true
                    data:
                      message: Logged out
                  summary: Logout Response
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
  /api/auth/refresh:
    post:
      operationId: auth_refresh
      summary: Refresh access token
      tags:
      - Authentication
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
            examples:
              RefreshRequest:
                value:
                  refresh: <jwt-refresh-token>
                summary: Refresh Request
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericSuccessEnvelope'
              examples:
                RefreshResponse:
                  value:
                    success: true
                    data:
                      access: <jwt-access-token>
                  summary: Refresh Response
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
          description: ''
...
components:
  schemas:
...
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
      required:
      - message
    ErrorEnvelope:
      type: object
      properties:
        success:
          type: boolean
          default: false
        error:
          $ref: '#/components/schemas/ErrorDetail'
      required:
      - error
...
    GenericSuccessEnvelope:
      type: object
      properties:
        success:
          type: boolean
          default: true
        data:
          type: object
          additionalProperties: {}
      required:
      - data
...
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 1
          maxLength: 150
        password:
          type: string
          minLength: 1
          maxLength: 128
      required:
      - password
      - username
    LogoutRequest:
      type: object
      properties:
        refresh:
          type: string
          minLength: 1
      required:
      - refresh
    RefreshRequest:
      type: object
      properties:
        refresh:
          type: string
          minLength: 1
      required:
      - refresh
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...